Move the DashboardDataStore into its own module scoped under @SiteScope
