+++ WooCommerce/src/main/kotlin/com/woocommerce/android/datastore/DataStoreModule.kt
@@ -3,7 +3,6 @@ package com.woocommerce.android.datastore
 import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.datastore.core.DataStoreFactory
-import androidx.datastore.dataStoreFile
 import androidx.datastore.preferences.core.PreferenceDataStoreFactory
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.preferencesDataStoreFile
@@ -11,11 +10,8 @@ import com.woocommerce.android.datastore.DataStoreType.ANALYTICS_CONFIGURATION
 import com.woocommerce.android.datastore.DataStoreType.ANALYTICS_UI_CACHE
 import com.woocommerce.android.datastore.DataStoreType.TRACKER
 import com.woocommerce.android.di.AppCoroutineScope
-import com.woocommerce.android.tools.SelectedSite
-import com.woocommerce.android.ui.dashboard.data.DashboardSerializer
 import com.woocommerce.android.ui.mystore.data.CustomDateRange
 import com.woocommerce.android.ui.mystore.data.CustomDateRangeSerializer
-import com.woocommerce.android.ui.mystore.data.DashboardDataModel
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -77,18 +73,5 @@ class DataStoreModule {
         },
         scope = CoroutineScope(appCoroutineScope.coroutineContext + Dispatchers.IO),
         serializer = CustomDateRangeSerializer
-    )
-
-    @Provides
-    fun provideDashboardDataStore(
-        appContext: Context,
-        @AppCoroutineScope appCoroutineScope: CoroutineScope,
-        site: SelectedSite
-    ): DataStore<DashboardDataModel> = DataStoreFactory.create(
-        produceFile = {
-            appContext.dataStoreFile("dashboard_configuration_${site.get().id}")
-        },
-        scope = CoroutineScope(appCoroutineScope.coroutineContext + Dispatchers.IO),
-        serializer = DashboardSerializer
     )
 }
