+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/dashboard/data/DashboardDataStore.kt
@@ -1,9 +1,12 @@
 package com.woocommerce.android.ui.dashboard.data
 
 import androidx.datastore.core.DataStore
+import com.woocommerce.android.di.SiteComponentEntryPoint
+import com.woocommerce.android.tools.SelectedSite
 import com.woocommerce.android.ui.mystore.data.DashboardDataModel
 import com.woocommerce.android.util.WooLog
 import com.woocommerce.android.util.WooLog.T
+import dagger.hilt.EntryPoints
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.map
@@ -11,8 +14,13 @@ import java.io.IOException
 import javax.inject.Inject
 
 class DashboardDataStore @Inject constructor(
-    private val dataStore: DataStore<DashboardDataModel>
+    selectedSite: SelectedSite
 ) {
+    private val dataStore: DataStore<DashboardDataModel> = EntryPoints.get(
+        selectedSite.siteComponent!!,
+        SiteComponentEntryPoint::class.java
+    ).dashboardDataStore()
+
     val dashboard: Flow<DashboardDataModel?> = dataStore.data
         .catch { exception ->
             // dataStore.data throws an IOException when an error is encountered when reading data
